Whatsapp Automater/requirements.txt
pywhatkit==5.0
tkcalendar==1.6.1
Pillow==8.4.0
